requirements.txt
Flask>=3.0
requests>=2.32
moviepy>=1.0.3
numpy>=1.26
Pillow>=10.0
srt>=3.5
pyttsx3>=2.90
scipy>=1.11
python-dotenv>=1.0
